restructure project files into components folder
